refactor(duckfn): 优化代码安全性注释和性能改进
- 为所有外部回调函数添加安全注释，明确指出由DuckDB保证参数有效性
- 在多个类型实现中添加clippy属性允许裸指针解引用，因为指针由DuckDB FFI提供
- 将数组容量计算从N as usize简化为N，提高性能
- 优化map和flat_map操作，使用更简洁的写法
- 将参数类型从Vec引用改为切片引用以提高灵活性
- 简化一些不必要的迭代器操作和条件判断
- 修复类型复杂度相关的警告并添加允许属性

+++ duckfn/src/value_types/duck_map.rs
@@ -4,6 +4,8 @@ use libduckdb_sys::duckdb_vector;
 use quack_rs::prelude::{ListVector, LogicalType, MapVector, TypeId, Value};
 use std::hash::Hash;
 
+// 裸指针由 DuckDB FFI 提供，此处直接解引用
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
 impl<K: DuckValueType + Hash + Eq, V: DuckValueType> DuckValueType for IndexMap<K, Option<V>> {
     fn type_id() -> TypeId {
         TypeId::Map
@@ -93,8 +95,8 @@ impl<K: DuckValueType + Hash + Eq, V: DuckValueType> DuckValueType for IndexMap<
     }
     fn read_by_duck_value_valid(value: &Value) -> DuckResult<Self> {
         let map_size = value.map_len();
-        let mut map = Self::with_capacity(map_size as usize);
-        for i in 0..map_size as usize {
+        let mut map = Self::with_capacity(map_size);
+        for i in 0..map_size {
             let k_value_option = value.map_key(i);
             let k = if let Some(k_value) = k_value_option {
                 K::read_by_duck_value(&k_value)?
@@ -121,6 +123,8 @@ trait Helper {
 impl<K: DuckValueType + Hash + Eq, V: DuckValueType> Helper for IndexMap<K, V> {
     type H = IndexMap<K, Option<V>>;
 }
+// 裸指针由 DuckDB FFI 提供，此处直接解引用
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
 impl<K: DuckValueType + Hash + Eq, V: DuckValueType> DuckValueType for IndexMap<K, V> {
     fn type_id() -> TypeId {
         <Self as Helper>::H::type_id()
@@ -199,8 +203,8 @@ impl<K: DuckValueType + Hash + Eq, V: DuckValueType> DuckValueType for IndexMap<
     }
     fn read_by_duck_value_valid(value: &Value) -> DuckResult<Self> {
         let map_size = value.map_len();
-        let mut map = Self::with_capacity(map_size as usize);
-        for i in 0..map_size as usize {
+        let mut map = Self::with_capacity(map_size);
+        for i in 0..map_size {
             let k_value_option = value.map_key(i);
             let k = if let Some(k_value) = k_value_option {
                 K::read_by_duck_value(&k_value)?
